Optimize final image size and rebase on full UBI (#17)
This change aims at optimizing the final image size.

The first thing it does is to switch to `ubi_/ubi-micro` base
image. This reduces the final image size from 273 MB to 77 MB.

The second thing is to use the `-ldflags="-s -w"` option of
the Go compiler to strip to binary. This reduces the size from
77 MB to 66 MB. In total, the image size is divided by 4.

This change also switches to the `ubi8/go-toolset` base image for the
builder stage. This is the image we recommend to the community and we
don't need all the OpenShift stuff to build a simple Go binary.

Signed-off-by: Fabien Dupont <fabiendupont@pm.me>